Backend/requirements.txt
Flask
Flask-Cors
Flask-SocketIO
python-socketio
python-engineio
Werkzeug
PyJWT
